requirements.txt
python-dateutil==2.8.2
streamlit==1.22.0